move safe_aggregate_metrics to utils.utils
